refactor: rename modules and files for consistency and clarity
