+++ .env
@@ -1,7 +1,7 @@
-LOG_CONFIG_FILE=src/config/logging.json
+LOG_CONFIG_FILE=src/config/log_config.json
 ESM_CHECKPOINTS_DIR=src/models
-ESM2_REPRESENTATION_CONFIG_FILE=src/config/esm2_representations.json
-AMINO_ACID_DESCRIPTORS_FILE=src/config/amino_acid_descriptors.csv
-FEATURE_COLLECTION_FILE=src/config/features_collection.json
-AD_METHODS_COLLECTION_FILE=src/config/ad_methods_collection.json
-OUTPUT_SETTINGS=src/config/output_settings.json
+ESM2_REPRESENTATION_CONFIG_FILE=src/config/esm2_representations_config.json
+AMINO_ACID_DESCRIPTORS_FILE=src/config/amino_acid_descriptors_config.csv
+FEATURES_CONFIG_FILE=src/config/features_config.json
+AD_METHODS_CONFIG_FILE=src/config/ad_methods_config.json
+OUTPUTS_CONFIG=src/config/outputs_config.json
