+++ src/config/outputs_config.json
@@ -1,5 +1,5 @@
 {
-  "output_settings": [
+  "outputs_config": [
     {
       "key": "checkpoints_path",
       "name": "Checkpoints",
